Lisp,refactor: export reusable code for implementing parsers for lisp-family
